fix: change suggestions
